pyproject.toml
[tool.poetry]
name = "get_and_process_data_massela"
version = "0.1.0"
description = "'Script responsavel por coleta e processar os dados'"
authors = ["Dieinimy Maganha <dieinimy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.1.41"
pandas = "^1.4.4"
requests = "^2.28.1"
SQLAlchemy = "^1.4.23"
psycopg2-binary = "^2.9.3"
notebook = "^6.4.12"
dynaconf = "^3.1.9"
xlrd = "^2.0.1"
python-dotenv = "^0.21.0"


[tool.poetry.dev-dependencies]
pep8 = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
